requirements.txt
fastapi
uvicorn
transformers
torch
httpx
python-dotenv